[FIX] Renamed HomeView and ProfileOverviewParent, removed console.logs
